requirements.txt
# FastAPI and related packages
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Security packages
python-jose[cryptography]>=3.3.0  # JWT token handling
passlib[bcrypt]>=1.7.4  # Password hashing
python-multipart>=0.0.6  # For form data handling
slowapi>=0.1.8  # Rate limiting for FastAPI

# Database
pymongo>=4.4.0
motor>=3.2.0  # Async MongoDB driver

# Vector Database
pinecone>=7.0.0
# weaviate-client>=3.18.0  # Uncomment if using Weaviate instead

# LangChain and related packages
langchain>=0.0.217
langchain-openai>=0.0.2

# LLM Providers
openai>=1.3.0
requests>=2.28.0  # Used for DeepSeek API integration

# HTTP Client
httpx>=0.24.0  # For local embeddings service

# Utilities
tenacity>=8.2.2
tiktoken>=0.4.0
numpy>=1.24.0
pandas>=2.0.0
redis>=4.5.0  # For caching (optional)

# Testing
pytest>=7.3.1
httpx>=0.24.1  # For testing FastAPI

# Embeddings and ML
sentence-transformers>=2.2.2
torch>=2.0.0  # Apple Silicon optimized version
transformers>=4.30.0
accelerate>=0.20.0  # For model optimization
safetensors>=0.3.1  # Faster model loading

# Email validation for Pydantic
email-validator>=2.0.0
